+++ src/Phan/AST/Visitor/KindVisitorImplementation.php
@@ -346,6 +346,11 @@ abstract class KindVisitorImplementation implements KindVisitor
         return $this->visit($node);
     }
 
+    public function visitTypeUnion(Node $node)
+    {
+        return $this->visit($node);
+    }
+
     public function visitNullableType(Node $node)
     {
         return $this->visit($node);